# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Concert Tracker.",
        "title": "Concert Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/like": {
            "get": {
                "description": "Returns the number of likes for a specific post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get number of likes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get number of likes",
                        "name": "userPostID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserLikeGetResponse"
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Toggles whether a user likes a given post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Like or unlike a post",
                "parameters": [
                    {
                        "description": "Like Toggle Payload",
                        "name": "like",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserLikePostPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like status toggled successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/userpost": {
            "get": {
                "description": "Gets public posts from a user's followed network, sorted with most recent first",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get posts for user dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of logged in user",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page number of posts (sets of 20)",
                        "name": "p",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity from user's followed network",
                        "schema": {
                            "$ref": "#/definitions/types.UserPostGetResponse"
                        }
                    },
                    "400": {
                        "description": "Error describing failure",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a post for a user. Can be set to public or private with IsPublic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create user post",
                "parameters": [
                    {
                        "description": "User Post Creation Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserPostCreatePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error describing failure - including duplicate attendance posts",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.UserLikeGetResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "types.UserLikePostPayload": {
            "type": "object",
            "required": [
                "userID",
                "userPostID"
            ],
            "properties": {
                "userID": {
                    "type": "integer"
                },
                "userPostID": {
                    "type": "integer"
                }
            }
        },
...
        "types.UserPostCreatePayload": {
            "type": "object",
            "required": [
                "authorID",
                "concertID",
                "isPublic",
                "type"
            ],
            "properties": {
                "authorID": {
                    "type": "integer"
                },
                "concertID": {
                    "type": "integer"
                },
                "isPublic": {
                    "type": "boolean"
                },
                "rating": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "ATTENDED",
                        "WISHLIST",
                        "REVIEW",
                        "LISTCREATED"
                    ]
                },
                "userPostID": {
                    "type": "integer"
                }
            }
        },
        "types.UserPostGetResponse": {
            "type": "object",
            "properties": {
                "authorID": {
                    "type": "integer"
                },
                "concertID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isPublic": {
                    "type": "boolean"
                },
                "rating": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userPostID": {
                    "type": "integer"
                }
            }
        },
...
    }
}